pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "qwen-image-edit-macos"
version = "1.0.0"
description = "Professional image generation and editing with Qwen Image Edit, optimized for Apple Silicon"
readme = "README.md"
license = {text = "Apache-2.0"}
authors = [
    {name = "Qwen Image Edit macOS", email = "hello@example.com"},
]
keywords = ["image-generation", "image-editing", "ai", "apple-silicon", "macos", "qwen", "diffusion"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Multimedia :: Graphics :: Graphics Conversion",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0", 
    "diffusers>=0.35.0",
    "transformers>=4.35.0",
    "accelerate>=0.25.0",
    "safetensors>=0.4.0",
    "pillow>=9.0.0",
    "numpy>=1.21.0",
    "tqdm>=4.65.0",
    "click>=8.0.0",
    "rich>=13.0.0",
    "pyyaml>=6.0.0",
    "requests>=2.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
qwen-image = "qwen_image_edit.cli.main:main"

[project.urls]
Homepage = "https://github.com/yourusername/qwen-image-edit-macos"
Repository = "https://github.com/yourusername/qwen-image-edit-macos"
Documentation = "https://github.com/yourusername/qwen-image-edit-macos#readme"
Issues = "https://github.com/yourusername/qwen-image-edit-macos/issues"

[tool.setuptools.packages.find]
include = ["qwen_image_edit*"]

[tool.black]
line-length = 88
target-version = ["py310", "py311"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
